+++ ErsatzTV.Application/MediaItems/Queries/GetMediaItemInfoHandler.cs
@@ -1,5 +1,6 @@
 using ErsatzTV.Core;
 using ErsatzTV.Core.Domain;
+using ErsatzTV.Core.Errors;
 using ErsatzTV.Core.Extensions;
 using ErsatzTV.Infrastructure.Data;
 using ErsatzTV.Infrastructure.Extensions;
@@ -40,10 +41,37 @@ public class GetMediaItemInfoHandler : IRequestHandler<GetMediaItemInfo, Either<
                 .ThenInclude(mv => mv.Streams)
                 .Include(i => (i as Episode).EpisodeMetadata)
                 .ThenInclude(mv => mv.Subtitles)
+                .Include(i => (i as Episode).Season)
+                .ThenInclude(s => s.Show)
+                .ThenInclude(s => s.ShowMetadata)
+                .Include(i => (i as OtherVideo).OtherVideoMetadata)
+                .ThenInclude(mv => mv.Subtitles)
+                .Include(i => (i as OtherVideo).MediaVersions)
+                .ThenInclude(mv => mv.Chapters)
+                .Include(i => (i as OtherVideo).MediaVersions)
+                .ThenInclude(mv => mv.Streams)
+                .Include(i => (i as Image).ImageMetadata)
+                .ThenInclude(mv => mv.Subtitles)
+                .Include(i => (i as Image).MediaVersions)
+                .ThenInclude(mv => mv.Chapters)
+                .Include(i => (i as Image).MediaVersions)
+                .ThenInclude(mv => mv.Streams)
+                .Include(i => (i as Song).SongMetadata)
+                .ThenInclude(mv => mv.Subtitles)
+                .Include(i => (i as Song).MediaVersions)
+                .ThenInclude(mv => mv.Chapters)
+                .Include(i => (i as Song).MediaVersions)
+                .ThenInclude(mv => mv.Streams)
+                .Include(i => (i as MusicVideo).MusicVideoMetadata)
+                .ThenInclude(mv => mv.Subtitles)
+                .Include(i => (i as MusicVideo).MediaVersions)
+                .ThenInclude(mv => mv.Chapters)
+                .Include(i => (i as MusicVideo).MediaVersions)
+                .ThenInclude(mv => mv.Streams)
                 .SelectOneAsync(i => i.Id, i => i.Id == request.Id)
                 .MapT(Project);
 
-            return mediaItem.ToEither(BaseError.New("Unable to locate media item"));
+            return mediaItem.ToEither<BaseError>(new UnableToLocateMediaItem());
         }
         catch (Exception ex)
         {
@@ -53,6 +81,8 @@ public class GetMediaItemInfoHandler : IRequestHandler<GetMediaItemInfo, Either<
 
     private static MediaItemInfo Project(MediaItem mediaItem)
     {
+        string displayTitle = Playouts.Mapper.GetDisplayTitle(mediaItem, Option<string>.None);
+
         MediaVersion version = mediaItem.GetHeadVersion();
 
         string serverName = mediaItem.LibraryPath.Library.MediaSource switch
@@ -68,7 +98,9 @@ public class GetMediaItemInfoHandler : IRequestHandler<GetMediaItemInfo, Either<
         List<Subtitle> subtitles = mediaItem switch
         {
             Movie m => m.MovieMetadata.Map(mm => mm.Subtitles).Flatten().ToList(),
-            Episode e => e.EpisodeMetadata.Map(mm => mm.Subtitles).Flatten().ToList(),
+            Episode e => e.EpisodeMetadata.Map(em => em.Subtitles).Flatten().ToList(),
+            MusicVideo mv => mv.MusicVideoMetadata.Map(mvm => mvm.Subtitles).Flatten().ToList(),
+            Song s => s.SongMetadata.Map(sm => sm.Subtitles).Flatten().ToList(),
             _ => []
         };
 
@@ -96,6 +128,7 @@ public class GetMediaItemInfoHandler : IRequestHandler<GetMediaItemInfo, Either<
 
         return new MediaItemInfo(
             mediaItem.Id,
+            displayTitle,
             mediaItem.GetType().Name,
             mediaItem.LibraryPath.Library.GetType().Name,
             serverName,